peimage: Copy the image header and ensure it's not clobbered
grub finds the embedded modules by finding the `mods` section
via the PE header. We did not copy the header, so grub found
no modules when we executed it.

shim solves the problem by copying the header, and checking
that the relocations physical addresses don't clobber it.

edk2 just reads the header into the target buffer, but does
no override checks.

We are extra cool people and hence check that virtual and
physical address of sections we copy do not clobber the
header.